Apply suggestions from code review
I see in other places that you use relative links, if this is OK it will make it easier for me to reuse the Markdown from this repo in the BYOC docs. But, I can just edit the URLs in the build script as I import the files if this causes an issue.